requirements.txt
fastapi
uvicorn[standard]

sentence-transformers
faiss-cpu

psycopg2-binary
python-dotenv
pydantic
pydantic-settings

numpy
pytest
